coverage: Extract function metadata handling to a `covfun` submodule
